Revert "Merge dev into Test"
This reverts commit 31fcb65673d871abe2beb2fa6268ff0a7ece4dae, reversing
changes made to 39303183f86ae472787055573889cf83d6952041.

+++ ada-payment-module/connectors/adapay.py
@@ -11,7 +11,7 @@ def lovelace_to_ada(amount):
 
 
 def ada_to_lovelace(amount):
-    return float(amount)*RATE
+    return int(amount)*RATE
 
 
 class AdaPayUrl(Enum):
